requirements.txt
pandas
numpy
langchain==0.0.335
openai==0.27.10
dill
aiolimiter==1.1.0
Flask==3.0.0
python-dotenv
tiktoken
psycopg2-binary
SQLAlchemy==2.0.20
